pyproject.toml
[project]
name = "rag-movie-plots"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.6"
dependencies = [
    "chromadb>=1.3.4",
    "datasets>=4.4.1",
    "evaluate>=0.4.6",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "langchain>=1.0.4",
    "langchain-chroma>=1.0.0",
    "langchain-huggingface>=1.0.1",
    "langchain-openai>=1.0.3",
    "langchain-text-splitters>=1.0.0",
    "matplotlib>=3.10.7",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
    "ragas>=0.3.8",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]
